Renaming things to sterling-patern.
