--- sterling/setup.cfg
@@ -1,4 +1,0 @@
-[develop]
-script_dir=$base/lib/sterling
-[install]
-install_scripts=$base/lib/sterling
