+++ sterling_patern_costmaps/alias.sh
@@ -6,7 +6,7 @@ alias run_gazebo_high='ros2 launch sterling_gazebo sidewalks.launch.py high_res:
 
 ### Step 2: Launch Nav2
 # Sterling parameters
-alias run_nav2='ros2 launch utexas_panther bringup_launch.py namespace:=panther observation_topic:=ouster/scan observation_topic_type:=laserscan slam:=true use_rviz:=True use_sim_time:=true params_file:=/root/ros2_ws/src/sterling/config/nav2_params.yaml'
+alias run_nav2='ros2 launch utexas_panther bringup_launch.py namespace:=panther observation_topic:=ouster/scan observation_topic_type:=laserscan slam:=true use_rviz:=True use_sim_time:=true params_file:=/root/ros2_ws/src/sterling_patern_costmaps/config/nav2_params.yaml'
 
 # How to send a goal_pose now and the alias does not work
 function run_goal() {
@@ -16,6 +16,6 @@ function run_goal() {
 }
 
 ### Step 3: Sterling costmap node
-alias run_sterling_costmap='source /root/ros2_ws/install/setup.bash && ros2 launch sterling costmaps.launch.py'
+alias run_sterling_patern_costmap='source /root/ros2_ws/install/setup.bash && ros2 launch sterling_patern_costmaps costmaps.launch.py'
 # While node is running, call service to save the costmap
-alias save_costmap='source /root/ros2_ws/install/setup.bash && ros2 service call /sterling/save_costmap std_srvs/srv/Trigger'
+alias save_costmap='source /root/ros2_ws/install/setup.bash && ros2 service call /sterling_patern_costmaps/save_costmap std_srvs/srv/Trigger'
