+++ sterling_patern_costmaps/config/nav2_params.yaml
