+++ sterling_patern_costmaps/config/panther_sim.rviz
