+++ sterling_patern_costmaps/config/params.yaml
@@ -5,7 +5,7 @@ panther/sterling: # Make sure namespace matches
       sub_topic_local_costmap: "/panther/local_costmap/costmap"
       pub_topic_local_costmap: "/panther/sterling/local_costmap"
       pub_topic_local_costmap_hz: 1.0
-      model_path: "/root/ros2_ws/src/sterling/models/patern_plus"
+      model_path: "/root/ros2_ws/src/sterling_patern_costmaps/models/patern_plus"
       # True, if you want to run the adapted Patern model. This looks for _adapted.pt files in the model folder.
       adapted: True
       # True, if you want to allow model to label obstacles from terrain
