+++ sterling_patern_costmaps/launch/costmaps.launch.py
@@ -14,8 +14,8 @@ from launch_ros.actions import Node
 
 
 def generate_launch_description():
-    params_file = os.path.join(get_package_share_directory("sterling"), "config", "params.yaml")
-    sterling_dir = get_package_share_directory("sterling")
+    params_file = os.path.join(get_package_share_directory("sterling_patern_costmaps"), "config", "params.yaml")
+    sterling_patern_costmaps_dir = get_package_share_directory("sterling_patern_costmaps")
 
     namespace = LaunchConfiguration("namespace")
     use_sim_time = LaunchConfiguration("use_sim_time")
@@ -40,7 +40,7 @@ def generate_launch_description():
     )
     declare_rviz_config_file_arg = DeclareLaunchArgument(
         "rviz_config_file",
-        default_value=os.path.join(sterling_dir, "config", "panther_sim.rviz"),
+        default_value=os.path.join(sterling_patern_costmaps_dir, "config", "panther_sim.rviz"),
         description="Full path to the RVIZ config file to use",
     )
 
@@ -51,14 +51,14 @@ def generate_launch_description():
             declare_use_rviz_arg,
             declare_rviz_config_file_arg,
             Node(
-                package="sterling",
+                package="sterling_patern_costmaps",
                 executable="local_costmap_builder",
                 name="local_costmap_builder",
                 namespace=namespace,
                 parameters=[params_file],
             ),
             Node(
-                package="sterling",
+                package="sterling_patern_costmaps",
                 executable="global_costmap_builder",
                 name="global_costmap_builder",
                 namespace=namespace,
