+++ sterling_patern_costmaps/package.xml
@@ -1,10 +1,10 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>sterling</name>
-  <version>0.0.0</version>
-  <description>TODO: Package description</description>
-  <maintainer email="nick.chan@utexas.edu">nchan</maintainer>
+  <name>sterling_patern_costmaps</name>
+  <version>0.1.0</version>
+  <description>Package for converting sterling/patern output into ROS costmaps</description>
+  <maintainer email="asharp@utexas.edu">asharp</maintainer>
   <license>TODO: License declaration</license>
 
   <depend>rclpy</depend>
