+++ sterling_patern_costmaps/record_bag_sim.sh
