+++ sterling_patern_costmaps/setup.cfg
@@ -1,0 +1,4 @@
+[develop]
+script_dir=$base/lib/sterling_patern_costmaps
+[install]
+install_scripts=$base/lib/sterling_patern_costmaps
