+++ sterling_patern_costmaps/sterling_patern_costmaps/nodes/global_costmap_builder.py
