+++ sterling_patern_costmaps/test/test_copyright.py
